+++ packages/contracts/src/gateway/post/request/index.ts
@@ -1,2 +1,3 @@
 export * from './post.request.js';
 export * from './admin-post.request.js';
+export * from './comment.request.js';