+++ packages/domain-post/src/models/index.ts
@@ -1,2 +1,3 @@
 export * from './post.model.js';
+export * from './comment.model.js';
 export * from './mapper.js';