+++ packages/post-processors/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 3.0.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/messaging@2.7.3
+  - @volontariapp/database@3.3.4
+  - @volontariapp/outbox@0.9.22
+
 ## 3.0.3
 
 ### Patch Changes